try using deathscreen.js for death in main battle room
